requirements.txt
olefile==0.44
paho-mqtt==1.3.1
Pillow==4.3.0
PyAutoGUI==0.9.36
PyMsgBox==1.0.6
PyScreeze==0.1.13
python3-xlib==0.15
PyTweening==1.0.3
